refactor(homepage, components): Create Hero component and replace on homepage
